requirements.txt
tqdm==4.64.1
spacy==3.4.2
pillow==9.3.0
spacy-transformers==1.1.8